+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './interval';
+export * from './timeout';
 export * from './types.types';